Add files via upload
After opening all of the m files in the folder, you should write the hours’ number in front of “iteration” in the “Main NSGAII” and “Test Case”. Then run the “Main NSGAII”, write the number 10 in the command window and press enter. The Algorithm will run, and it shows you the Pareto front. After that, you can write the “iteration” in the “separating_Pop” and find one of the solutions that is the best solution. The numbers in the first row of the “ans” matrix are decision variables and the second row shows the objective functions.